+++ src/main/java/ir/najaftech/gui/Contacts/FormEvent.java
@@ -1,4 +1,4 @@
-package ir.najaftech.gui;
+package ir.najaftech.gui.Contacts;
 
 import java.util.EventObject;
