+++ src/main/java/ir/najaftech/gui/Contacts/NumericTextFieldEnforcer.java
@@ -1,4 +1,4 @@
-package ir.najaftech.gui;
+package ir.najaftech.gui.Contacts;
 
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
