+++ src/main/java/ir/najaftech/gui/Contacts/TextPanel.java
@@ -1,4 +1,4 @@
-package ir.najaftech.gui;
+package ir.najaftech.gui.Contacts;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
